refactor(layout): consolidate global providers into unified component
- Remove app/[language]/Providers.tsx
- Create widgets/layout/ui/global-providers.tsx with enhanced structure
- Simplify root layout by extracting provider logic
- Update global error and not-found pages to use new provider
- Add timeZone configuration to NextIntlClientProvider
- Maintain all existing functionality while improving code organization

This change improves maintainability by centralizing provider configuration
and makes the layout structure more consistent across different page types.
