+++ app/global-error.tsx
@@ -4,6 +4,7 @@ import { hasLocale, Locale } from 'next-intl';
 import { useParams } from 'next/navigation';
 import NotExist from '@/shared/components/NotExist';
 import { routing } from '@/shared/i18n/routing';
+import GlobalProviders from '@/widgets/layout/ui/global-providers';
 import './global.css';
 
 function GlobalErrorPage() {
@@ -13,12 +14,17 @@ function GlobalErrorPage() {
     ? params?.language
     : routing.defaultLocale;
 
+  const head = (
+    <>
+      <title>未知錯誤 | 島島阿學</title>
+      <link rel="shortcut icon" href="/assets/brand/favicon.png" />
+    </>
+  );
+
   return (
-    <html lang={locale}>
-      <body>
-        <NotExist />
-      </body>
-    </html>
+    <GlobalProviders head={head} locale={locale}>
+      <NotExist />
+    </GlobalProviders>
   );
 }
