+++ app/global-not-found.tsx
@@ -4,6 +4,7 @@ import { hasLocale, Locale } from 'next-intl';
 import { useParams } from 'next/navigation';
 import NotExist from '@/shared/components/NotExist';
 import { routing } from '@/shared/i18n/routing';
+import GlobalProviders from '@/widgets/layout/ui/global-providers';
 import './global.css';
 
 function GlobalNotFoundPage() {
@@ -13,12 +14,17 @@ function GlobalNotFoundPage() {
     ? params?.language
     : routing.defaultLocale;
 
+  const head = (
+    <>
+      <title>找不到頁面 | 島島阿學</title>
+      <link rel="shortcut icon" href="/assets/brand/favicon.png" />
+    </>
+  );
+
   return (
-    <html lang={locale}>
-      <body>
-        <NotExist />
-      </body>
-    </html>
+    <GlobalProviders head={head} locale={locale}>
+      <NotExist />
+    </GlobalProviders>
   );
 }
